+++ src/quacc/recipes/tblite/phonons.py
@@ -23,6 +23,9 @@ def phonon_flow(
     atoms: Atoms,
     symprec: float = 1e-4,
     min_lengths: float | tuple[float, float, float] | None = 20.0,
+    supercell_matrix: (
+        tuple[tuple[int, int, int], tuple[int, int, int], tuple[int, int, int]] | None
+    ) = None,
     displacement: float = 0.01,
     t_step: float = 10,
     t_min: float = 0,
@@ -54,6 +57,9 @@ def phonon_flow(
         Precision for symmetry detection.
     min_lengths
         Minimum length of each lattice dimension (A).
+    supercell_matrix
+        The supercell matrix to use. If specified, it will override any
+        value specified by `min_lengths`.
     displacement
         Atomic displacement (A).
     t_step
@@ -91,6 +97,7 @@ def phonon_flow(
         relax_job=relax_job_ if run_relax else None,
         symprec=symprec,
         min_lengths=min_lengths,
+        supercell_matrix=supercell_matrix,
         displacement=displacement,
         t_step=t_step,
         t_min=t_min,