+++ tests/core/recipes/emt_recipes/test_emt_phonons.py
@@ -20,6 +20,14 @@ def test_phonon_flow(tmp_path, monkeypatch):
     assert Path(tmp_path / "phonopy.yaml").is_file()
     assert Path(tmp_path, "phonopy_auto_band_structure.yaml").is_file()
 
+    atoms = bulk("Cu")
+    output = phonon_flow(atoms, supercell_matrix=((2, 0, 0), (0, 2, 0), (0, 0, 2)))
+    assert output["results"]["thermal_properties"]["temperatures"].shape == (101,)
+    assert output["results"]["thermal_properties"]["temperatures"][0] == 0
+    assert output["results"]["thermal_properties"]["temperatures"][-1] == 1000
+    assert output["results"]["force_constants"].shape == (8, 8, 3, 3)
+    assert "mesh_properties" in output["results"]
+
 
 def test_phonon_flow_v2(tmp_path, monkeypatch):
     monkeypatch.chdir(tmp_path)